+++ features/merchant/review_inbox/src/main/java/com/tokopedia/review/feature/inbox/buyerreview/data/pojo/inboxdetail/ReviewImageUrl.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class ReviewImageUrl(
     @SerializedName("attachment_id")
     @Expose
-    val attachmentId: Int = 0,
+    val attachmentId: Long = 0,
 
     @SerializedName("description")
     @Expose
